refactor lai cac folder comment
