+++ src/app/app.module.ts
@@ -21,7 +21,7 @@ import {CreateCommentComponent} from './component/comments/comment-create/create
 import { EditCommentComponent } from './component/comments/comment-edit/edit-comment.component';
 import { PostEditComponent } from './component/post/post-edit/post-edit.component';
 import { PostDeleteComponent } from './component/post/post-delete/post-delete.component';
-import { DeleteCommentComponent } from './component/comments/delete-comment/delete-comment.component';
+import { DeleteCommentComponent } from './component/comments/comment-delete/delete-comment.component';
 
 @NgModule({
   declarations: [
