+++ src/app/component/comments/comment-list/list-comment.component.html
@@ -3,6 +3,6 @@
   {{comment.user.username}}
   <br>
   {{comment.description}}
-  <button [routerLink]="['/comments/edit-comment/',comment.id]">Edit</button>
-  <button [routerLink]="['/comments/delete-comment/',comment.id]">Delete</button>
+  <button [routerLink]="['/comments-edit/',comment.id]">Edit</button>
+  <button [routerLink]="['/comments-delete/',comment.id]">Delete</button>
 </div>
